afxdp/cos/queue_service: relocate inline tests to tests.rs
queue_service/ is already a directory module (mod.rs + drain.rs +
service.rs). Mirrors the umem/, coordinator/, forwarding/, frame/,
session_glue/ split patterns: drop the inline test block from mod.rs
into a sibling tests.rs.

Mechanical changes:
- Inline mod tests block (lines 1137-2309) extracted to sibling
  tests.rs (1,176 lines incl 5-line header)
- 4-space dedent applied uniformly
- mod.rs ends with:
    #[cfg(test)]
    #[path = "tests.rs"]
    mod tests;
- pre-rustfmt'd

mod.rs drops from 2,310 → 1,139 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
